Create spritebouncer7.bas
Trying to get rid of the F(0)=128 array ... and find a memory location to peek from instead, saving precious bytes
So i tried to find the sequence of 128,0,1 in memory, preferably at a lower address,. 
The sequence was not found anywhere in memory.

Ok so i scanned the memory for instances of the sequence 0,1.
Found it at a lower address (677), so i poked 676,128 and crossed fingers, that it would not break anything. Just a nice Ready prompt, which means i can get rid of the end stuff, and shorten the poke, but is getting a penalty with 3 chars (128), still it saves 1 more byte.